+++ packages/ente-ui/src/cookie.ts
@@ -1,7 +1,15 @@
-import { Maybe } from "monet";
-import * as cookie from "js-cookie";
+import { Maybe, Some, None } from "monet";
 
 export const get = (key: string): Maybe<string> => {
-  const c = cookie.get(key);
-  return !!c ? Maybe.Some(c) : Maybe.Nothing();
+  const value = "; " + document.cookie;
+  const parts = value.split("; " + key + "=");
+
+  if (parts.length === 2) {
+    const result = parts
+      .pop()
+      .split(";")
+      .shift();
+    return Some(result);
+  }
+  return None();
 };